updated how functions work and improved functionality
